+++ tests/data/forms/flat/form.b.ts
@@ -6,6 +6,7 @@ import svkExtend from "../../extension/svk";
 
 import dvr from "../../../../src/validators/DVR";
 import svk from "../../../../src/validators/SVK";
+import { ValidationPlugins } from "../../../../src/models/ValidatorInterface";
 
 const fields = {
   username: {
@@ -33,7 +34,7 @@ const schema = {
   },
 };
 
-const plugins = {
+const plugins: ValidationPlugins = {
   dvr: dvr(validatorjs),
   svk: svk({
     schema,