+++ tests/data/forms/flat/form.i.ts
@@ -3,6 +3,7 @@ import { Form } from "../../../../src";
 import svk from "../../../../src/validators/SVK";
 import { FormInterface } from "../../../../src/models/FormInterface";
 import { OptionsModel } from "../../../../src/models/OptionsModel";
+import { ValidationPlugins } from "../../../../src/models/ValidatorInterface";
 
 const fields = {
   username: {
@@ -28,7 +29,7 @@ const schema = {
   },
 };
 
-const plugins = {
+const plugins: ValidationPlugins = {
   svk: svk({
     package: ajv,
     schema,