+++ tests/data/forms/flat/form.m.ts
@@ -8,6 +8,7 @@ import dvr from "../../../../src/validators/DVR";
 import svk from "../../../../src/validators/SVK";
 import { FormInterface } from "../../../../src/models/FormInterface";
 import { OptionsModel } from "../../../../src/models/OptionsModel";
+import { ValidationPlugins } from "../../../../src/models/ValidatorInterface";
 
 const fields = {
   username: {
@@ -35,7 +36,7 @@ const schema = {
   },
 };
 
-const plugins = {
+const plugins: ValidationPlugins = {
   svk: svk({
     package: ajv,
     schema,